% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_intermed.R
\name{ci.intermed}
\alias{ci.intermed}
\title{Confidence intervals for interventional mediation effects}
\usage{
\method{ci}{intermed}(object, est = "aiptw", level = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{"intermed"}}

\item{est}{The estimate to obtain CI around. Options are \code{"tmle"} and \code{"aiptw"}}

\item{level}{Level of the confidence interval}

\item{...}{Other options (ignored)}
}
\description{
Confidence intervals for interventional mediation effects
}
