% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_Q_M}
\alias{estimate_Q_M}
\title{estimate_Q_M}
\usage{
estimate_Q_M(
  A,
  M1,
  M2,
  C,
  DeltaA,
  DeltaM,
  SL_Q_M,
  glm_Q_M = NULL,
  a_0,
  stratify,
  verbose = FALSE,
  return_models = FALSE,
  valid_rows,
  all_mediator_values,
  return_list_by_a_0 = FALSE,
  ...
)
}
\arguments{
\item{A}{A vector of binary treatment assignment (assumed to be equal to 0 or
1).}

\item{M1}{A \code{vector} of mediators.}

\item{M2}{A \code{vector} of mediators.}

\item{C}{A \code{data.frame} of named covariates.}

\item{DeltaA}{Indicator of missing treatment (assumed to be equal to 0 if
missing 1 if observed).}

\item{DeltaM}{Indicator of missing outcome (assumed to be equal to 0 if
missing 1 if observed).}

\item{glm_Q_M}{A character describing a formula to be used in the call to
\code{glm} for the outcome regression.}

\item{a_0}{A list of fixed treatment values}

\item{stratify}{A \code{boolean} indicating whether to estimate the outcome
regression separately for observations with \code{A} equal to 0/1 (if
\code{TRUE}) or to pool across \code{A} (if \code{FALSE}).}

\item{verbose}{A boolean indicating whether to print status updates.}

\item{return_models}{A boolean indicating whether to return model fits for the
outcome regression, propensity score, and reduced-dimension regressions.}

\item{valid_rows}{A \code{list} of length \code{cvFolds} containing the row
indexes of observations to include in validation fold.}

\item{all_mediator_values}{All combinations of M1 and M2}

\item{return_list_by_a_0}{For power users, return the list prior to reformatting}

\item{...}{Additional arguments (not currently used)}

\item{SL_Q}{A vector of characters or a list describing the Super Learner
library to be used for the outcome regression.}

\item{family}{A character passed to \code{SuperLearner}}
}
\description{
Helper function to estimate the mediator distribution. 
Returns an n-length list, where each entry is a 2-length list
corresponding to mediator distributions under each treatment assignment.
Within each of these is another list where there are three entries corresponding
to the bivariate distribution, and each marginal distribution.
}
\details{
The bivariate distribution is estimated by estimating the conditional
distribution of M1 given A, C, and M2 and the marginal distribution of M1 given
A and C. In each case, we use a hazard-based estimation approach for estimating
these distributions. The
}
