% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_Qbar}
\alias{estimate_Qbar}
\title{estimate_Qbar}
\usage{
estimate_Qbar(Y, A, M1, M2, C, DeltaA, DeltaM, SL_Q, glm_Q = NULL, a_0,
  stratify, family, verbose = FALSE, return_models = FALSE, valid_rows,
  ...)
}
\arguments{
\item{Y}{A vector of continuous or binary outcomes.}

\item{A}{A vector of binary treatment assignment (assumed to be equal to 0 or
1).}

\item{M1}{A \code{vector} of mediators.}

\item{M2}{A \code{vector} of mediators.}

\item{C}{A \code{data.frame} of named covariates.}

\item{DeltaA}{Indicator of missing treatment (assumed to be equal to 0 if
missing 1 if observed).}

\item{DeltaM}{Indicator of missing outcome (assumed to be equal to 0 if
missing 1 if observed).}

\item{SL_Q}{A vector of characters or a list describing the Super Learner
library to be used for the outcome regression.}

\item{glm_Q}{A character describing a formula to be used in the call to
\code{glm} for the outcome regression.}

\item{a_0}{A list of fixed treatment values}

\item{stratify}{A \code{boolean} indicating whether to estimate the outcome
regression separately for observations with \code{A} equal to 0/1 (if
\code{TRUE}) or to pool across \code{A} (if \code{FALSE}).}

\item{family}{A character passed to \code{SuperLearner}}

\item{verbose}{A boolean indicating whether to print status updates.}

\item{return_models}{A boolean indicating whether to return model fits for the
outcome regression, propensity score, and reduced-dimension regressions.}

\item{valid_rows}{A \code{list} of length \code{cvFolds} containing the row
indexes of observations to include in validation fold.}

\item{...}{Additional arguments (not currently used)}

\item{all_mediator_values}{All combinations of M1 and M2}
}
\description{
Function to estimate outcome regression as a function of \code{A, C, M1,} 
and \code{M2}. Because later we will need to marginalize these estimates over
estimated distributions of \code{M1} and \code{M2}, the output includes the predicted 
value for each C_i, i = 1, ..., n and for every value of \code{A} in 
\code{a_0}. The output is formatted as an n-length list where there is one entry
for each observation. This entry includes a list of predicted values under each treatment
\code{Qbar_a_0}, which is itself a list with a vector of predictions for each value of 
\code{a_0}. Also included is an entry called \code{which_M1_obs}, which indicates rows of
the \code{all_mediator_values} that correspond to this observation's observed value of 
\code{M1} and \code{M2}. Similarly, there is a vector \code{which_M2_obs}, and also a vector
\code{which_M1_M2_obs}, which indicates the row of \code{all_mediator_values} that 
corresponds to this observation's observed value of BOTH \code{M1} and \code{M2}.
}
