% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_eif.R
\name{evaluate_direct_effect}
\alias{evaluate_direct_effect}
\title{Helper function to evaluate the direct effect}
\usage{
evaluate_direct_effect(Qbarbar, use_conditional = TRUE)
}
\description{
Helper function to evaluate the direct effect
}
