% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_eif.R
\name{evaluate_eif_direct}
\alias{evaluate_eif_direct}
\title{Function to evaluate the canonical gradient of 
the direct effect at a given set of estimated nuisance parameters
and for each observation.}
\usage{
evaluate_eif_direct(Y, A, M1, M2, Qbar, Q_M, Qbarbar, gn, a, a_star,
  use_conditional, ...)
}
\arguments{
\item{Y}{The outcome}

\item{A}{The treatment}

\item{Qbar}{Outcome regression list}

\item{Q_M}{Mediator distribution list}

\item{Qbarbar}{A list; needs to have entries named M1_M2_a and M1_star_M2_star_a_star corresponding
to, respectively, the outcome regression under A = a, marginalized with respect to joint 
mediator distribution given C and A = a, and the outcome regression under A = a_star, marginalized
with respect to the joint mediator distribution given C and A = a_star.}

\item{gn}{List of propensity scores}

\item{a}{Treatment value}

\item{a_star}{Referent treatment value}

\item{use_conditional}{Boolean. Should the conditional_total_effect be used (generally used for when evaluating
EIF of targeted nuisance parameters, since there we target directly the difference)}
}
\description{
Function to evaluate the canonical gradient of 
the direct effect at a given set of estimated nuisance parameters
and for each observation.
}
