% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_eif.R
\name{evaluate_eif_indirect_M2}
\alias{evaluate_eif_indirect_M2}
\title{Function to evaluate the canonical gradient of 
the indirect effect through M2 at a given set of estimated nuisance parameters
and for each observation.}
\usage{
evaluate_eif_indirect_M2(Y, A, M1, M2, Qbar, Q_M, Qbarbar, gn, a, a_star,
  all_mediator_values, ...)
}
\arguments{
\item{Y}{The outcome}

\item{A}{The treatment}

\item{M1}{First mediator}

\item{M2}{Second mediator}

\item{Qbar}{Outcome regression list}

\item{Q_M}{Mediator distribution list}

\item{Qbarbar}{A list; needs to have entries named M1_M2_a and M1_star_M2_star_a_star corresponding
to, respectively, the outcome regression under A = a, marginalized with respect to joint 
mediator distribution given C and A = a, and the outcome regression under A = a_star, marginalized
with respect to the joint mediator distribution given C and A = a_star.}

\item{gn}{List of mediator values}

\item{a}{Treatment value}

\item{a_star}{Referent treatment value}
}
\description{
Function to evaluate the canonical gradient of 
the indirect effect through M2 at a given set of estimated nuisance parameters
and for each observation.
}
