% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_eif.R
\name{evaluate_eif_total}
\alias{evaluate_eif_total}
\title{Function to evaluate the canonical gradient of 
the total effect at a given set of estimated nuisance parameters
and for each observation.}
\usage{
evaluate_eif_total(Y, A, Qbarbar, gn, a, a_star, ...)
}
\arguments{
\item{Y}{The outcome}

\item{A}{The treatment}

\item{Qbarbar}{A list; needs to have entries named M1_M2_a and M1_star_M2_star_a_star corresponding
to, respectively, the outcome regression under A = a, marginalized with respect to joint 
mediator distribution given C and A = a, and the outcome regression under A = a_star, marginalized
with respect to the joint mediator distribution given C and A = a_star.}

\item{gn}{List of mediator values}

\item{a}{Treatment value}

\item{a_star}{Referent treatment value}
}
\description{
Function to evaluate the canonical gradient of 
the total effect at a given set of estimated nuisance parameters
and for each observation.
}
