% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_eif.R
\name{evaluate_total_effect}
\alias{evaluate_total_effect}
\title{Helper function to evaluate the total effect}
\usage{
evaluate_total_effect(Qbarbar)
}
\description{
Helper function to evaluate the total effect
}
