% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_joint}
\alias{extract_joint}
\title{Helper function for extracting joint distributions at particular values of 
the mediator.}
\usage{
extract_joint(Q_M_n_i, M1_i, M2_i, all_mediator_values, a_val = "a_star", ...)
}
\arguments{
\item{Q_M_n_i}{An entry in the \code{Q_M_n} list, where the first entry in the list
corresponds to mediator distributions under \code{a_star}; the second entry in the list
corresponds to mediator distributions under \code{a}. In each of these lists, we have
the joint of M1,M2; the marginal of M1; and the marginal of M2.}

\item{M1_i}{The particular value of \code{M1} at which you are interested in evaluating the joint density.
Note that the general idea is to \code{mapply} over this function to extract the joint
under each observed value of \code{C} and extract the marginal distribution at the corresponding
value of \code{M1} and \code{M2}.}

\item{all_mediator_values}{All possible combinations of \code{M1} and \code{M2}. Used to identify
which entry in \code{Q_M_n_i} that corresponds to values specified in \code{M1_i} and \code{M2_i}.}

\item{a_val}{Which treatment are you interested in extracting the marginal under? Possible values
are \code{"a_star"} and \code{"a"}}

\item{...}{Not used}
}
\description{
Helper function for extracting joint distributions at particular values of 
the mediator.
}
