% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_marginal}
\alias{extract_marginal}
\title{Helper function for extracting marginal distributions at particular values of 
the mediator.}
\usage{
extract_marginal(Q_M_n_i, mediator = "M1", M_i, unique_M_values,
  a_val = "a_star", ...)
}
\arguments{
\item{Q_M_n_i}{An entry in the \code{Q_M_n} list, where the first entry in the list
corresponds to mediator distributions under \code{a_star}; the second entry in the list
corresponds to mediator distributions under \code{a}. In each of these lists, we have
the joint of M1,M2; the marginal of M1; and the marginal of M2.}

\item{mediator}{Which mediator are you interested in extracting the marginal of? Possible
values are \code{"M1"} and \code{"M2"}.}

\item{M_i}{The particular value at which you are interested in evaluating the marginal density.
Note that the general idea is to \code{mapply} over this function to extract the marginal
under each observed value of \code{C} and extract the marginal distribution at the corresponding
value of e.g., \code{M1}.}

\item{unique_M_values}{The unique values that the selected \code{mediator} can assume. Recall that
the ordering of the marginal entries in \code{Q_M_n_i} is in this order.}

\item{a_val}{Which treatment are you interested in extracting the marginal under? Possible values
are \code{"a_star"} and \code{"a"}}

\item{...}{Not used}
}
\description{
Helper function for extracting marginal distributions at particular values of 
the mediator.
}
