% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_Qbar_obs}
\alias{extract_Qbar_obs}
\title{Helper function for extracting Qbar at the observed values of confounders and
mediators under \code{a} or \code{a_star}}
\usage{
extract_Qbar_obs(Qbar_n_i, M1_i, M2_i, all_mediator_values, a_val = "a_star")
}
\description{
Helper function for extracting Qbar at the observed values of confounders and
mediators under \code{a} or \code{a_star}
}
