% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{get_Qbarbar}
\alias{get_Qbarbar}
\title{Helper function to marginalize outcome regression over mediator 
distributions. Used to get each of the relevant nuisance parameters needed
to evaluate the total, direct, and indirect effects.}
\usage{
get_Qbarbar(Qbar_n, Q_M_n, all_mediator_values, ...)
}
\description{
Helper function to marginalize outcome regression over mediator 
distributions. Used to get each of the relevant nuisance parameters needed
to evaluate the total, direct, and indirect effects.
}
