% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_intermed.R
\docType{data}
\name{n}
\alias{n}
\title{Function to compute TMLE of intervention effects}
\format{An object of class \code{numeric} of length 1.}
\usage{
n
}
\arguments{
\item{Y}{outcome}

\item{C}{confounders}

\item{M1}{first mediator}

\item{M2}{second mediator}

\item{A}{treatment}

\item{a}{value of A of interest}

\item{a_star}{other value of A of interest}

\item{glm_Q_M}{A list with named element \code{M1} and \code{M2} describing the right-hand side
of a GLM formula for, respectively, the conditional hazard of M1 given A (if stratify = FALSE),
C, and M2 and the conditional hazard of M2 given A (if stratify = FALSE) and C. The formula's 
can additionally include a variable named \code{bin_id}, which corresponds to the numeric bin number
(i.e., this controls the level of smoothing over the range of \code{M1} and \code{M2}).}
}
\description{
Function to compute TMLE of intervention effects
}
\keyword{datasets}
