% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predict_density}
\alias{predict_density}
\title{Helper function to predict from a super learner fit on the hazard scale
to obtain a prediction on the density scale}
\usage{
predict_density(
  sl_fit_conditional,
  sl_fit_marginal,
  how_predict,
  all_mediator_values,
  a_val,
  validC,
  M1,
  M2,
  valid_n,
  stratify
)
}
\arguments{
\item{sl_fit_conditional}{A fitted SuperLearner object for the M1 | M2, C, A regression}

\item{sl_fit_marginal}{A fitted SuperLearner object for the M2 | C, A regression}

\item{how_predict}{Needed to figure out how to predict from model. Possible choices
are "SuperLearner", "single_algo", and "glm"}

\item{all_mediator_values}{data.frame of all observed mediator values}

\item{a_val}{Relevant treatment level}

\item{validC}{Confounders in the validation set}

\item{M1}{All M1 values}

\item{M2}{All M2 values}

\item{valid_n}{Number of obs in validation set}

\item{stratify}{Were sl_fit's stratified?}
}
\description{
Helper function to predict from a super learner fit on the hazard scale
to obtain a prediction on the density scale
}
