% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reduce_merge}
\alias{reduce_merge}
\title{Helper to merge covariates in targeting}
\usage{
reduce_merge(...)
}
\arguments{
\item{...}{Arguments passed in}
}
\description{
Helper to merge covariates in targeting
}
