% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{reorder_list}
\alias{reorder_list}
\title{Helper function to reorder lists according to cvFolds}
\usage{
reorder_list(a_list, a_0, valid_rows, n_SL = 1, which_nuisance, n,
  n_mediator_values, n_M1_values = 0, n_M2_values = 0)
}
\arguments{
\item{a_list}{Structured list of nuisance parameters}

\item{a_0}{Treatment levels}

\item{valid_rows}{List of rows of data in validation data for
each split.}

\item{n_SL}{Number of super learners. If >1, then predictions
are averaged over repeated super learners}

\item{which_nuisance}{Structure of reordering is slightly different for 
each relevant nuisance parameter ("OR", "PS", "MED")}

\item{n}{Sample size}

\item{n_mediator_values}{Number of unique combinations of mediators}
}
\description{
Helper function to reorder lists according to cvFolds
}
