% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{target_conditional_total_effect}
\alias{target_conditional_total_effect}
\title{The naming convention is a bit different here, because we're actually
able to go after the conditional total effect directly. That is, 
we can define a loss function whose minimizer defines the conditional mean of the
difference between Qbar_a
and Qbar_a_star with respect to the joint distribution of M1 and M2 given C and A.
We can then define a submodel through this conditional mean difference (which is 
exactly the conditional total effect) and target this quantity directly.}
\usage{
target_conditional_total_effect(Qbarbar, Qbar, Y, A, a, a_star, ...)
}
\arguments{
\item{Y}{The outcome}

\item{A}{The treatment}

\item{a}{The comparison value of treatment}

\item{a_star}{The referent value of treatment}
}
\description{
The naming convention is a bit different here, because we're actually
able to go after the conditional total effect directly. That is, 
we can define a loss function whose minimizer defines the conditional mean of the
difference between Qbar_a
and Qbar_a_star with respect to the joint distribution of M1 and M2 given C and A.
We can then define a submodel through this conditional mean difference (which is 
exactly the conditional total effect) and target this quantity directly.
}
