% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{target_Qbar}
\alias{target_Qbar}
\title{Function to target the outcome regression}
\usage{
target_Qbar(Y, A, M1, M2, a, a_star, all_mediator_values, Qbar_n, gn,
  Q_M_n, ...)
}
\arguments{
\item{Y}{the outcome}

\item{A}{the observed treatment}

\item{Qbar_n}{The n-length list created as a result of the outcome
regression estimation routine}

\item{gn}{The 2-length list created as a result of the propensity 
score estimation routine}

\item{Q_M_n}{The n-length list created as a result of the mediator
distribution estimation routine}
}
\description{
Function to target the outcome regression
}
