% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_intermed.R
\name{intermed}
\alias{intermed}
\title{Function to compute TMLE of intervention effects}
\usage{
intermed(Y, C, M1, M2, A, DeltaA = as.numeric(!is.na(A)),
  DeltaM = as.numeric(!is.na(M1)), a = 1, a_star = 0,
  SL_Qbar = NULL, SL_g = NULL, SL_Q_M = NULL, n_SL = 1,
  glm_Qbar = NULL, glm_g = NULL, glm_Q_M = NULL, tolg = 0.01,
  tol = 1/(sqrt(length(Y)) * log(length(Y))), targeted_se = FALSE,
  return_models = FALSE, cvFolds = 1, parallel = FALSE,
  use_future = FALSE, Qbar_n = NULL, Q_M_n = NULL, gn = NULL,
  max_iter = 50, verbose = FALSE, stratify = FALSE, ...)
}
\arguments{
\item{Y}{outcome}

\item{C}{confounders}

\item{M1}{first mediator}

\item{M2}{second mediator}

\item{A}{treatment}

\item{a}{value of A of interest}

\item{a_star}{other value of A of interest}

\item{glm_Q_M}{A list with named element \code{M1} and \code{M2} describing the right-hand side
of a GLM formula for, respectively, the conditional hazard of M1 given A (if stratify = FALSE),
C, and M2 and the conditional hazard of M2 given A (if stratify = FALSE) and C. The formula's 
can additionally include a variable named \code{bin_id}, which corresponds to the numeric bin number
(i.e., this controls the level of smoothing over the range of \code{M1} and \code{M2}).}
}
\description{
Function to compute TMLE of intervention effects
}
