% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{target_conditional_total_effect}
\alias{target_conditional_total_effect}
\title{The naming convention is a bit different here, because we're actually
able to go after the conditional total effect directly. That is, 
we can define a loss function whose minimizer defines the conditional mean of the
difference between Qbar_a
and Qbar_a_star with respect to the joint distribution of M1 and M2 given C and A.
We can then define a submodel through this conditional mean difference (which is 
exactly the conditional total effect) and target this quantity directly.}
\usage{
target_conditional_total_effect(Qbarbar, gn, Qbar, Y, A, a, a_star, ...)
}
\arguments{
\item{Qbarbar}{Iterated mean estimates}

\item{gn}{Power users may wish to pass in their own properly formatted list of the
propensity score so that
nuisance parameters can be fitted outside of \code{intermed}.}

\item{Qbar}{Outcome regression estimates}

\item{Y}{A vector of continuous or binary outcomes.}

\item{A}{A vector of binary treatment assignment (assumed to be equal to 0 or
1).}

\item{a}{The label for the treatment. The effects estimates returned pertain
to estimation of interventional effects of \code{a} versus \code{a_star}.}

\item{a_star}{The label for the treatment. The effects estimates returned pertain
to estimation of interventional effects of \code{a} versus \code{a_star}.}
}
\description{
The naming convention is a bit different here, because we're actually
able to go after the conditional total effect directly. That is, 
we can define a loss function whose minimizer defines the conditional mean of the
difference between Qbar_a
and Qbar_a_star with respect to the joint distribution of M1 and M2 given C and A.
We can then define a submodel through this conditional mean difference (which is 
exactly the conditional total effect) and target this quantity directly.
}
