% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{target_Qbarbar_M1_star_times_M2_star_a}
\alias{target_Qbarbar_M1_star_times_M2_star_a}
\title{Function for targeting Qbarbar_M1_star_times_M2_star_a}
\usage{
target_Qbarbar_M1_star_times_M2_star_a(
  Qbarbar,
  Y,
  A,
  a,
  a_star,
  gn,
  tol = 1/(sqrt(length(Y)) * log(length(Y))),
  ...
)
}
\arguments{
\item{Qbarbar}{List of marginalized parameters}

\item{Y}{The outcome}

\item{A}{The treatment}

\item{a}{The comparison value of treatment}

\item{a_star}{The referent value of treatment}

\item{tol}{Tolerance for iterative TMLE}
}
\description{
There are two interesting features of this 
targeting problem. First, we see that the nuisance parameter Qbarbar_M1_star_times_M2_star_a
can be viewed in two ways: (1) the conditional mean of Qbarbar_M1_star_a given C with respect
to the marginal of M_2 given A = a_star, C; (2) the conditional mean of Qbarbar_M2_star_a given C 
with respect to the marginal of M_1 given A = a_star, C. The natural inclination then 
is to use a sum loss function. Here it looks like we actually can use a sum 
loss approach, so long as the IPTW are incorporated into the loss function. We make
two copies of each observation with A = a_star; assign Qbarbar_M1_star_a as outcome in
half and Qbarbar_M2_star_a in the other half; then do one-shot targeting
}
